Code and preferences changed to record and set extrude speeds in mm/minute as a double.  This works for extruders that understand this, but existing extruders with DC motors running in open-loop just keep whatever PWM number they were using (usually about 215) and set that value in the extruder speed preferences file as if that were mm/minute.  The only perceptible difference should be that the extruder speeds in the preferences file have been renamed with the suffix "(mm/minute)" when they used to have "(0..255)".

+++ reprap/host/src/org/reprap/devices/GCodeExtruder.java
@@ -9,7 +9,7 @@ public class GCodeExtruder extends GenericExtruder
 {
 	GCodeReaderAndWriter gcode;
 	
-	int currentSpeed = 0;
+	double currentSpeed = 0;
 	
 	/**
 	 * @param prefs
@@ -34,7 +34,7 @@ public class GCodeExtruder extends GenericExtruder
 		return Double.parseDouble(gcode.queueRespond("M105; get temperature").substring(2)); // Throw away "T:"
 	}
 	
-	public void setExtrusion(int speed, boolean reverse) throws IOException
+	public void setExtrusion(double speed, boolean reverse) throws IOException
 	{
 		if(extrusionSpeed < 0)
 			return;
